web/requirements.txt
Django==4.0.3
gunicorn==20.1.0
psycopg2==2.8.6
requests==2.25.1
pytz==2021.1
Pillow==8.2.0
django-ckeditor==6.1.0
pydub==0.25.1
ffmpeg==1.4
django-phonenumber-field==6.1.0
phonenumbers==8.12.44